Fix CSRF in WebAuthn support (#5304)
