+++ support/cas-server-support-thymeleaf/src/main/resources/templates/fragments/webAuthnLogin.html
@@ -122,7 +122,11 @@
             <input type="hidden" name="execution" th:value="${flowExecutionKey}"/>
         </form>
 
+        <input type="hidden" id="loginCsrfToken" th:name="${_csrf.parameterName}" th:value="${_csrf.token}"/>
+
         <script type="text/javascript">
+            const csrfToken = $('#loginCsrfToken').val();
+
             let div = document.querySelector('#divDeviceInfo');
             new mdc.dataTable.MDCDataTable(div);
