+++ support/cas-server-support-webauthn-core-webflow/src/main/java/org/apereo/cas/webauthn/web/flow/WebAuthnMultifactorWebflowConfigurer.java
@@ -3,16 +3,20 @@ package org.apereo.cas.webauthn.web.flow;
 import org.apereo.cas.configuration.CasConfigurationProperties;
 import org.apereo.cas.util.CollectionUtils;
 import org.apereo.cas.web.flow.CasWebflowConstants;
+import org.apereo.cas.web.flow.actions.ConsumerExecutionAction;
 import org.apereo.cas.web.flow.configurer.AbstractCasMultifactorWebflowConfigurer;
 import org.apereo.cas.web.flow.configurer.CasMultifactorWebflowCustomizer;
+import org.apereo.cas.web.support.WebUtils;
 import org.apereo.cas.webauthn.WebAuthnCredential;
 
 import lombok.val;
 import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.security.web.csrf.CsrfTokenRepository;
 import org.springframework.util.StringUtils;
 import org.springframework.webflow.definition.registry.FlowDefinitionRegistry;
 import org.springframework.webflow.engine.builder.support.FlowBuilderServices;
 
+import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -31,18 +35,35 @@ public class WebAuthnMultifactorWebflowConfigurer extends AbstractCasMultifactor
 
     private static final String TRANSITION_ID_VALIDATE_WEBAUTHN = "validateWebAuthn";
 
+    private final CsrfTokenRepository csrfTokenRepository;
+
     public WebAuthnMultifactorWebflowConfigurer(final FlowBuilderServices flowBuilderServices,
                                                 final FlowDefinitionRegistry loginFlowDefinitionRegistry,
                                                 final FlowDefinitionRegistry flowDefinitionRegistry,
                                                 final ConfigurableApplicationContext applicationContext,
                                                 final CasConfigurationProperties casProperties,
-                                                final List<CasMultifactorWebflowCustomizer> mfaFlowCustomizers) {
+                                                final List<CasMultifactorWebflowCustomizer> mfaFlowCustomizers,
+                                                final CsrfTokenRepository csrfTokenRepository) {
         super(flowBuilderServices, loginFlowDefinitionRegistry, applicationContext,
             casProperties, Optional.of(flowDefinitionRegistry), mfaFlowCustomizers);
+        this.csrfTokenRepository = csrfTokenRepository;
     }
 
     @Override
     protected void doInitialize() {
+        val addCsrfTokenAction = new ConsumerExecutionAction(context -> {
+            val request = WebUtils.getHttpServletRequestFromExternalWebflowContext(context);
+            val response = WebUtils.getHttpServletResponseFromExternalWebflowContext(context);
+            request.setAttribute(HttpServletResponse.class.getName(), response);
+
+            var csrfToken = csrfTokenRepository.loadToken(request);
+            if (csrfToken == null) {
+                csrfToken = csrfTokenRepository.generateToken(request);
+                csrfTokenRepository.saveToken(csrfToken, request, response);
+            }
+            context.getFlowScope().put(csrfToken.getParameterName(), csrfToken);
+        });
+
         multifactorAuthenticationFlowDefinitionRegistries.forEach(registry -> {
             val flow = getFlow(registry, MFA_WEB_AUTHN_EVENT_ID);
             createFlowVariable(flow, CasWebflowConstants.VAR_ID_CREDENTIAL, WebAuthnCredential.class);
@@ -66,6 +87,7 @@ public class WebAuthnMultifactorWebflowConfigurer extends AbstractCasMultifactor
             val viewRegState = createViewState(flow, "viewRegistrationWebAuthn", "webauthn/casWebAuthnRegistrationView");
             viewRegState.getEntryActionList().addAll(
                 createEvaluateAction(CasWebflowConstants.ACTION_ID_POPULATE_SECURITY_CONTEXT),
+                addCsrfTokenAction,
                 createEvaluateAction("webAuthnStartRegistrationAction"),
                 setPrincipalAction);
             createTransitionForState(viewRegState, CasWebflowConstants.TRANSITION_ID_SUBMIT, CasWebflowConstants.STATE_ID_SAVE_REGISTRATION);
@@ -79,7 +101,7 @@ public class WebAuthnMultifactorWebflowConfigurer extends AbstractCasMultifactor
             val viewLoginFormState = createViewState(flow,
                 CasWebflowConstants.STATE_ID_VIEW_LOGIN_FORM, "webauthn/casWebAuthnLoginView", loginBinder);
             createStateModelBinding(viewLoginFormState, CasWebflowConstants.VAR_ID_CREDENTIAL, WebAuthnCredential.class);
-            viewLoginFormState.getEntryActionList().addAll(
+            viewLoginFormState.getEntryActionList().addAll(addCsrfTokenAction,
                 createEvaluateAction("webAuthnStartAuthenticationAction"), setPrincipalAction);
             createTransitionForState(viewLoginFormState, TRANSITION_ID_VALIDATE_WEBAUTHN,
                 CasWebflowConstants.STATE_ID_REAL_SUBMIT, Map.of("bind", Boolean.TRUE, "validate", Boolean.TRUE));
@@ -96,12 +118,15 @@ public class WebAuthnMultifactorWebflowConfigurer extends AbstractCasMultifactor
 
         val flow = getLoginFlow();
         if (flow != null && webAuthn.getCore().isAllowPrimaryAuthentication()) {
+            val appId = org.apache.commons.lang3.StringUtils.defaultString(webAuthn.getCore().getApplicationId(), casProperties.getServer().getName());
             val setAppIdAction = createSetAction("flowScope." + WebAuthnStartRegistrationAction.FLOW_SCOPE_WEB_AUTHN_APPLICATION_ID,
-                StringUtils.quote(webAuthn.getCore().getApplicationId()));
+                StringUtils.quote(appId));
             flow.getStartActionList().add(setAppIdAction);
 
             val setPrimaryAuthAction = createSetAction("flowScope.webAuthnPrimaryAuthenticationEnabled", "true");
             flow.getStartActionList().add(setPrimaryAuthAction);
+
+            flow.getStartActionList().add(addCsrfTokenAction);
 
             val viewLoginFormState = getState(flow, CasWebflowConstants.STATE_ID_VIEW_LOGIN_FORM);
             createTransitionForState(viewLoginFormState, TRANSITION_ID_VALIDATE_WEBAUTHN, "validateWebAuthnToken");
