+++ support/cas-server-support-webauthn-core-webflow/src/main/java/org/apereo/cas/webauthn/web/flow/WebAuthnStartRegistrationAction.java
@@ -6,16 +6,12 @@ import org.apereo.cas.web.flow.actions.AbstractMultifactorAuthenticationAction;
 import org.apereo.cas.web.support.WebUtils;
 import org.apereo.cas.webauthn.WebAuthnMultifactorAuthenticationProvider;
 
-import com.yubico.core.RegistrationStorage;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
-import org.springframework.security.web.csrf.CsrfTokenRepository;
 import org.springframework.webflow.execution.Event;
 import org.springframework.webflow.execution.RequestContext;
-
-import javax.servlet.http.HttpServletResponse;
 
 /**
  * This is {@link WebAuthnStartRegistrationAction}.
@@ -33,11 +29,7 @@ public class WebAuthnStartRegistrationAction extends AbstractMultifactorAuthenti
      */
     public static final String FLOW_SCOPE_WEB_AUTHN_APPLICATION_ID = "webauthnApplicationId";
 
-    private final RegistrationStorage webAuthnCredentialRepository;
-
     private final CasConfigurationProperties casProperties;
-
-    private final CsrfTokenRepository csrfTokenRepository;
 
     @Override
     protected Event doExecute(final RequestContext requestContext) {
@@ -55,17 +47,6 @@ public class WebAuthnStartRegistrationAction extends AbstractMultifactorAuthenti
             flowScope.put("displayName", principal.getId());
         }
         flowScope.put(FLOW_SCOPE_WEB_AUTHN_APPLICATION_ID, webAuthn.getApplicationId());
-
-        val request = WebUtils.getHttpServletRequestFromExternalWebflowContext(requestContext);
-        val response = WebUtils.getHttpServletResponseFromExternalWebflowContext(requestContext);
-        request.setAttribute(HttpServletResponse.class.getName(), response);
-
-        var csrfToken = csrfTokenRepository.loadToken(request);
-        if (csrfToken == null) {
-            csrfToken = csrfTokenRepository.generateToken(request);
-            csrfTokenRepository.saveToken(csrfToken, request, response);
-        }
-        flowScope.put(csrfToken.getParameterName(), csrfToken);
         return null;
     }
 }
