+++ support/cas-server-support-webauthn/src/test/java/org/apereo/cas/webauthn/web/flow/WebAuthnStartRegistrationActionTests.java
@@ -74,7 +74,6 @@ public class WebAuthnStartRegistrationActionTests {
         assertNull(webAuthnStartRegistrationAction.execute(context));
         assertTrue(context.getFlowScope().contains(WebAuthnStartRegistrationAction.FLOW_SCOPE_WEB_AUTHN_APPLICATION_ID));
         assertTrue(context.getFlowScope().contains("displayName"));
-        assertTrue(context.getFlowScope().contains("_csrf"));
 
         val http = new HttpSecurity(mock(ObjectPostProcessor.class),
             new AuthenticationManagerBuilder(mock(ObjectPostProcessor.class)),
